Chore: FormationBoard 컴포넌트명 변경
